Generalise dired fontification
There is no need to limit it to the denote-directory.  We ultimately
want to fontify all Denote-style file names, not just the notes created
by Denote.

For example, I have been recording all my longer-term storage using such
a naming scheme: this mode gives me the extra faces for .pdf, .mp3, and
other files.

Users of Denote may want this for attachments.

The upside of having this as a buffer-local mode is that the user can
write a wrapper function that applies the mode only in a given
directory (like we were doing before).

Thanks to Ypot for suggesting a kernel of this idea in issue 1 over at
the GitHub mirror: <https://github.com/protesilaos/denote/issues/1>.