Add denote-dired-rename-file
This is not limited to notes that were created with Denote: it works on
any file in any directory.  The idea is to apply the Denote-style file
name in more contexts, such as for longer-term storage and attachments
to notes.

Also see commit 431124f, which is thematically aligned with this one.

Thanks to Ypot for giving me the idea in issue 1 over at the GitHub
mirror: <https://github.com/protesilaos/denote/issues/1>.