BREAKING remove support for hardcoded backlinks
We will find a more robust solution that will be with us for the
long-term.  The gist is that we need links that work even when the user
renames the underlying file.  Only the identifier is a constant: it is
unreasonable to expect that, say, the keywords in the file name of a
note will never change.  As such, writing the full path is not a viable
option.

Relevant discussions on the GitHub mirror:

* <https://github.com/protesilaos/denote/issues/13>
* <https://github.com/protesilaos/denote/issues/13>

This, by the way, is the first time a project of mine is busy on GitHub.
You are welcome to use the mirror or the official mailing list.
Whatever works for you (that's why I maintain all those sources, after
all).