REMOVE denote-link-register-ol-hyperlink variable
It was never a good idea to make this optional because it will make the
logic of buttonizing links much more complicated.  Furthermore, links in
Org also need to do the right thing during exporting, so we cannot
merely rely on a simple buttonization---registering the custom hyperlink
type is a requirement in that regard.