Use xref instead of relying on find+grep
With xref we delegate to a built-in facility instead of passing a
hard-coded list of arguments to 'process-lines'.

The old method is prone to errors, such as not working on builds of
Emacs for macOS, as reported by Colin McLear in issue 18 over at the
GitHub mirror: <https://github.com/protesilaos/denote/issues/18>.

Users of Emacs 28 or higher can configure 'xref-search-program' to
change from the default 'grep' to 'ripgrep', 'ugrep', or a user-defined
alternative.

THIS CHANGE IS PROVISIONAL and subject to further edits.