Make denote-link-add-links use id: when necessary
This will account for the target file.  So if we insert files matching a
regexp, we get the appropriate link type per file.

If it is Org and 'denote-link-use-org-id' is non-nil, then the link type
is 'id:'.

If it is another file type the link is 'denote:'.

If 'denote-link-use-org-id' is nil, the link is also 'denote:'.