Fix issue in denote-retrieve.el and refactor the code.
When a note title ended with a ")" (or other special characters),
denote-retrieve--value-title did not return the last character.

The regexps have been modified to focus only on the key part of the
line. The value part is easy to retrieve once the key is matched. We
just need to trim (of spaces and quotes) the remainder of the line.

denote-retrieve--value has been removed. denote-retrieve--search is
doing all the work. Given a KEY-REGEXP, it retrieves the value part of
the matching line. With optional KEY argument, it retrieves the key
part instead.