Use better default for file dates in front matter
The old user option 'denote-front-matter-date-format' is deprecated.  It
is superseded by 'denote-date-format'.

Now we use an inactive timestamp for Org, the RFC3339 standard for
Markdown (YAML or TOML), and the ISO 8601 format for plain text.

The user can specify a string value for 'denote-date-format', which
contains format specifiers as described in the doc string of
'format-time-string'.

Thanks to Peter Prevos and Kaushal Modi for the discussion in issue 27
over at the GitHub mirror: <https://github.com/protesilaos/denote/issues/27>.