Add entry to FAQ about xref's search failures on Windows
Even though 'denote' itself does not require external programs, some
of Emacs' features require external programs.  The 'xref' package is
one of those features.

While some of xref's features do not need any external applications,
the function 'xref-matches-in-files' used in 'denote-retrieve--xrefs'
will use 'xref-search-program' and by that also the associative list
'xref-search-program-alist' to find the actual shell command.  As
`-in-files` the suffix implies, the shell command needs to understand
multiple file names, which is the reason for 'xargs' in the pipe.

Since there is no Windows native alternative for 'xargs', the user
needs to install one of the many packages like msys64, MinGW, or Git,
that provide the necessary tools within their installation.
Afterwards, the path has to be added to the ~%PATH%~ environment
variable.  Note that 'exec-path' is not enough, since the shell proxy
on Windows, 'cmdproxy.exe', does not know about that Lisp-only
variable.

Either setting the PATH variable within Emacs as shown in the
changeset or setting the PATH variable within Windows' settings should
suffice and therefore get rid of this issue.