Refine how 'denote:' Org link is loaded
This should load the setup only after Org is made available.  Whereas
what we had before would throw an error if the user was not already
using Org.

Thanks to Abin Simon for reporting the problem and for showing how
Elfeed handles this issue.  This was done in issue 47 on the GitHub
mirror: <https://github.com/protesilaos/denote/issues/47>.