Fix error of date parsing in Emacs 28
The idea is to turn 2020-01-15 into 2020-01-15 16:19 so that the hour
and minute component is not left to 00:00.

This reduces the burden on the user who would otherwise need to input
that value in order to avoid the error of duplicate identifiers.

It also addresses a difference between Emacs 28 and Emacs 29 where the
former does not read dates without a time component.

Thanks to Peter Prevos for the feedback in issue 58 over at the GitHub
mirror: <https://github.com/protesilaos/denote/issues/58>.