Rewrite all documentation about renaming files
Bring everything up to date and clarify relevant points.

NOTE that I kept the symbol 'denote-dired-rename-file' instead of
replacing it with 'denote-dired-rename-file-and-rewrite-front-matter'.
The obsoletion was done in pull request 51 over at the GitHub mirror but
I found the longer name harder to differentiate from the command
'denote-dired-rename-file-and-add-front-matter' (which is appropriately
named): <https://github.com/protesilaos/denote/pull/51>.