REFACTOR post-rename front matter rewrite
It should now update only the title and keywords, without erasing the
whole block and replacing it with a new one.

The old method involved (i) duplication of code and (ii) would overwrite
custom front matter that the user would maintain in their file.

The new method should be technically better as well as address the
latter problem.

Thanks to Peter Prevos for the feedback in issue 60 over at the GitHub
mirror: <https://github.com/protesilaos/denote/issues/60>.