Rework denote--filetype-heuristics
denote--filetype-heuristics should always return an md filetype if the
file extension is md even if no key regexps match in the file. In that
case, we use the first md type in denote-file-types.

markdown-yaml should appear before markdown-toml in denote-file-types
because it used to be the default markdown type of
denote--filetype-heuristics.