Use active region in default title prompt
Facilitate a slightly smoother note taking by passing the contents of
the region from the denote command to denote--title-prompt as the
DEFAULT-TITLE argument, when the region is active.

The idea behind this DWIM-flavored patch is to be able to take a note
about a subject that appears in a buffer by simply marking it before
invoking the denote command, which feels natural to me.