Add support for `org-store-link`.
Hi,

Denote, uses the `denote:` link type internally.

This patch adds support for `org-store-link` of denote files meaning
that you can call `org-store-link` from a note and get a `denote:` link
instead of the current `file:` link.

The main benefit of a denote link is that it is resilient to renames,
etc. A file link would brake if the note was ever to be renamed.

As a side benefit, as `org-capture` uses `org-store-link` now you can
capture tasks on notes with resilient links!

This may seem related to `denote-link-use-org-id` though that one would
only work with orgmode notes (and, I think, as long as those were
withing `org-directory` when using `id:` links).

This should work with notes in any format, now `org-capture` works and
`org-store-link` too.

Best,
Marc

PS: This closes #104 on Github[1].
PSS: Second attempt at submitting a patch via git-email ...

[1]: https://github.com/protesilaos/denote/issues/104

Signed-off-by: Marc Fargas <marc@marcfargas.com>