Make denote-directory-files-matching-regexp public
This is the same idea as with commit c294eaf.

It means that users or developers can rely on this function in their own
code.  By making it "public" (removing the double hyphens), we declare
that we will keep this form in check and document any changes to its
behaviour.