Add denote-file-prompt instead of private function
We are working towards version 1.0.0.  The idea is to provide functions
that users/developers can rely on.  By making them public (removing the
double hyphens) we signify that those forms are (i) in a stable
state, and (ii) any change to them will be handled with care and
documented accordingly.