Improve conditionality of Org button fontification
This way we only try to use the correct face when we are inside of a
Denote note.  We do not want Org to try to figure this out in other
situations, such as while building the agenda.

Thanks to Alan Schmitt for reporting the problem on the mailing list.
It led me down the path of debugging the issue:
<https://lists.sr.ht/~protesilaos/denote/%3C877d2e2ini.fsf%40m4x.org%3E>.