Make denote-extract-keywords-from-path public
We are working towards version 1.0.0.  The idea is to provide functions
that users/developers can rely on.  By making them public (removing the
double hyphens) we signify that those forms are (i) in a stable
state, and (ii) any change to them will be handled with care and
documented accordingly.

Thanks to Peter Prevos for informing me about this function on the
mailing list in relation to the 'denote-statistics' package:

- <https://github.com/pprevos/denote-statistics>
- <https://lists.sr.ht/~protesilaos/denote/%3C87o7vio7ix.fsf%40protesilaos.com%3E#%3C87r10cexe6.fsf@prevos.net%3E>