Move some commands from denote.el to the manual
It is easier for the maintenance of the project to have such code in the
manual.  Otherwise we are setting a precedent where we need to maintain
variants of the same command that have a slightly different behaviour.

Perhaps we will find a more elegant solution that gives the user maximum
flexibility.

This is discussed on the mailing list:
<https://lists.sr.ht/~protesilaos/denote/%3C87fsgvddny.fsf%40protesilaos.com%3E>.