Change how 'denote-file-types' :front-matter is defined
This will allow changing front-matter variables by 'setq' after
'denote' is loaded/compiled without needing to set 'define-file-types'
again.

Currently, changing variables 'denote-*-front-matter' to personalize
the front matter with using 'setq' requires the user to set variable
'denote-file-types' again to reflect the change.

This is because 'denote-file-types' is defined with macro expansion
with using the backquote ("`") and "," to evaluate variables
'denote-*-front-matter' inside the macro.  This results in the front
matters fixed by the values of front-matter variables at the time of
loading or compiling 'denote', and setting the front-matter variables
thereafter with 'setq' will not affect the front matter in notes
created -- this will require setting 'denote-file-types' again.

For the relevant discussion, read:
<https://lists.sr.ht/~protesilaos/denote/%3C86k05gsqsg.fsf%40nobiot.com%3E>.