Allow nil :front-matter in denote-file-types
The 'denote-file-types' is, for the time being, specified as a
'defvar' for advanced users.  We do eventually want to turn it into a
defcustom, assuming there is a need for it (we will figure it out
through further testing).

This change allows the ':front-matter' property to accept a nil value,
which we internally read as an empty string.

Thanks to Noboru Ota (nobiot) for pointing this out on the mailing list:
<https://lists.sr.ht/~protesilaos/denote/%3C86k05gsqsg.fsf%40nobiot.com%3E>.