feat: use region as link title on denote-link
It is a common use-case to create a link in existing notes. In such cases, some
text is already there, and user just want to create a link on that text (instead
of inserting the link's title). With this change, if a region is active when
user tries to insert a link, the region is replaced with the link while using
the region's content as link's title.