fix: org-mode-like insert-link behavior with empty active region
If user has whitespace-only region active when inserting a denote-link, behave
org-mode does when inserting a link in this case. i.e selected region is
deleted, and an id-only link is inserted. Active region is trimmed regardless,
so if user try to link "hello " part of "hello world", the text is changed to
"helloworld" with "hello" being the link description and "world" plain text.