Clarify denote-link region behaviour in the manual
Thanks to Charanjit Singh for the contribution, which was done in pull
request 110 on GitHub following our discussion:
<https://github.com/protesilaos/denote/pull/110>.

This, together with Charanjit's other contribution in commits b350035,
fe66333 is within the limits of what is allowed without assigning
copyright to the Free Software Foundation.  Though the author will
need to do the paperwork for any future contributions.