Make renaming remove keywords if empty
The idea is an empty "keywords" entry in the front matter should mean
that the "__KEYWORDS" component of the title should not be present.
We already allow this, such as if the 'denote-prompts' is configured
to not have a 'keywords' prompt (check its doc string).

THIS NEEDS MORE TESTING TO BE SURE THERE ARE NO REGRESSIONS.