Display context of identifier in backlinks buffer with xref
The intention of the changes below is to display context of identifier
in the files in the backlinks buffer.  Currently Denote's backlinks
buffer display a list of filenames.  Denote currently retrieves
xref-items (xrefs-alist) and then strips off the context information and
utilizes only the file name (group) for backlinks buffer.  We can utilize
the xref-items as retrieved and make use of the xref's built-in facility
to display the context each identifier is in.  This also lets the
backlinks buffer display multiple occurrences of the same identifier in a
single file (as treated by Xref).

* denote.el

(denote--retrieve-files-in-xrefs): Currently unused;
commented out.  If sorting of the xref-groups (filenames) is critical,
we will need to first look into xrefs-alist and check if it is not
already done as desired by Xref.  If not, will need to consider a
different way to sort xrefs-alist.

(denote--retrieve-process-grep): Changed to return xrefs-alist as
retrieved by 'denote--retrieve-xrefs'.  The file name of the current
buffer where 'denote-link-backlinks' is called from is deleted so that
only backlinks appear.

(denote-backlinks-mode-map): Buttons no longer used.  Navigation is done
in the same way as Xref.  Key "g" is redefined to call 'revert-buffer'
instead of 'xref-revert-buffer'.

(denote-backlinks-mode): The major mode is now derived from
'xref--xref-buffer-mode', which itself is derived from 'special-mode'.

(denote-link--prepare-backlinks): Instead of importing files, this function now works off xrefs-alist.  The alist
includes the filename and other information, including "summary" -- the line the identifier is included.  Summary may
be available only Emacs 28.1 and onward (refer to function 'xref-item-summary').  As Denote currently supports 27.2,
this may require some way to work around it. 27.2 has not been tested.

(denote-xref--insert-xrefs): This function is a slightly modified
version of the built-in function 'xref--insert-xrefs'.  The only change
is the use of function
'denote-get-file-name-relative-to-denote-directory' for the file name
(group in xref).

(denote-link-backlinks): Replace "files" with "xrefs-alist" to make
explicit the data type that is meant to be used in line with all the
other changes above.  In addition, 'xref--push-markers' is called so that
'xref-quit-and-pop-marker-stack' (bound to "M-," by default) will work.
This can display the note's buffer where 'denote-link-backlinks' was
called from.