Make denote-link-fontify-backlinks obsolete
It will confuse users now that we have another option for displaying
the context of links.  It is better to just handle this internally.

Furthermore, our fontification does not look nice when the context is
shown.  In that case, the standard Xref style looks better because
it (i) clearly distinguishes file names from the context, (2) line
numbers from the context, and (3) highlights the matching identifier
on the given line.

Read the discussion on the mailing list about the refactoring of the
backlinks' buffer, with the participation of Noboru Ota:
<https://lists.sr.ht/~protesilaos/denote/%3C86r0yvzm12.fsf%40nobiot.com%3E>.

Also check the recent commit log for the relevant changes, including
patches from Noboru.