Add wrapper motions for backlinks' buffer
With the denote-backlinks-show-context, we now have two types of
backlinks' buffer: (1) a generic one that uses buttons that we create
and (2) the familiar Xref buffer.  These commands are meant to
accommodate both types.

Read the discussion on the mailing list about the refactoring of the
backlinks' buffer, with the participation of Noboru Ota:
<https://lists.sr.ht/~protesilaos/denote/%3C86r0yvzm12.fsf%40nobiot.com%3E>.

Also check the recent commit log for the relevant changes, including
patches from Noboru.