Update minimum required Emacs version to 28.1
Emacs 27 lacks certain Xref facilities that we need for the
backlinking facility.  It also adds to the maintenance burden.

With the help of Noboru Ota, we realised that Denote was broken on
Emacs 27 for quite a while.[1] The fact that we received no feedback
about it suggests that this change is the best course of action going
forward.

[1] <https://lists.sr.ht/~protesilaos/denote/%3C86r0yvzm12.fsf%40nobiot.com%3E#%3C86sfja78ik.fsf@nobiot.com%3E>