Refactor backlink xref functions
The main intent of this change is to remove 'denote-xref--insert-xrefs',
which was meant to be a slight modification to the built-in function
'xref--insert-xrefs'.  This way, it is intended that future maintenance
work for Denote will be lighter and more in line with the changes to the
built-in Xref library.

Additionally, refactoring is done for 'denote--retrieve-xrefs'. The
intent was to fulfill requirements (A) (B) and (1) (2) described below
in a way that avoids duplication of work -- an attempt to follow the DRY
(Don't Repeat Yourself) principle.

The function name 'denote--retrieve-xrefs' indicates that it should be
generic that should return all the xrefs as retrieved by the Xref
function (A); however, the current usage is only to retrieve backlinks
for a specific file, by excluding a relevant element from the xref-alist
(B). This is used by two different commands: (1) 'denote-link-backlinks'
and (2) 'denote-link-find-backlink'; they have different requirements
for the output format: (1) requires relative file names and (2) requires
an absolute file names.

There can be possibly further refactoring.  That is now left to future iteration.

More detail of change to each altered function follows.

* denote.el

(denote--retrieve-xrefs): Add an optional argument FILE. It is meant to
cater to the different requirements of (A) and (B) above.

(denote--retrieve-process-grep): This function is now used only by
'denote-link-backlinks' and refactored to transform GROUP from absolute
file name to relative one for `denote-directory`. This is requirement
(1) above for display purposes of the backlinks' buffer As the doc
string of 'xref--insert-xrefs' notes that GROUP in xref-alist is for
"decoration purposes" and it is also processed in the same way in
'xref--analyze', this should not be a problem.

(denote-link-find-backlink): Refactor to avoid function
'(denote--retrieve-process-grep)' to retrieve xref-alist for requirement
(2) above.

(denote-link--prepare-backlinks): Refactor to use 'xref--insert-xrefs'
directly. As GROUP in xref-alist (GROUP is its key) now carries relative
file name for display purposes, the original logic to insert a list of
file names is also refactored to remove the duplicate work of function
'denote-get-file-name-relative-to-denote-directory'.  It is now done
centrally in 'denote--retrieve-process-grep' for the purpose of
backlinks' buffer (with or without the context).

(denote-xref--insert-xrefs): Remove this function completely.