Make denote-file-types read a :date-function
The end goal is to make this variable the single point of entry to all
file-type-aware operations.  This way, expert users can add support
for any file type they want.  Depending on how it goes, we may even
make it a user option.

Relevant thread on the GitHub mirror issue 86:
<https://github.com/protesilaos/denote/issues/86>.