Update Acknowledgements for commit 39a20a7
Thanks to Graham Marlow for the contribution which was done in pull
request 112 on the GitHub mirror:
<https://github.com/protesilaos/denote/pull/112>.

The contribution, with the tweaks made in 95cd9f8, is within the
eligible line count: it does not require copyright assignment to the
Free Software Foundation.  Though Graham would need to do the
paperwork for any future patch.