denote-org: reword Commentary
There are some Org-specific features in denote.el, like the 'denote:'
hyperlink type, which are useful for all Org users.  Whereas what we
have in denote-org.el is meant to be explicitly requested with a
'require' call and, thus, is more specialised in its
scope/functionality.

This does not make it any less important.  It just makes it easier for
us to organise our code.