feat: xref-backend facility to enable xref integration
Add xref-backend facility for Denote.

To enable this feature, `xref-backend-functions' like this.

    (add-hook 'xref-backend-functions #'denote--xref-backend)

This lets users use the standard Xref commands such as
'xref-find-definitions' ("M-." by default) 'xref-find-references' ("M-?"
by default) on an identifier in the body of text.  To use this, in a
note in a directory, move your point to an identifier like this and
simply invoke one of these commands.  'find-definition' should take you
to the note that identifier points to.  'find-references' by default
should open up a completion in minibuffer with a list of identifiers and
the one at point should already be selected as the default.  Simply
press RET to open a Xref buffer with all references of the identifier.

You can tell xref-references not to prompt by adding the following:

    (add-to-list 'xref-prompt-for-identifier #'xref-find-references
    :append)

Note that you do not need (1) to use any markup to explicitly indicate
the text string at point is a link or (2) to select a range of text.