add crosslink to Karl Voit's work
Denote started with management of thoughts/ideas in Emacs org-mode
and extends to files.  Some of the tools initiated by Karl Voit do
offer the addition of time stamps, key words from a controlled
vocabulary, etc on files (e.g., date2name) which may be processed
further (memacs).  It is worth to hint to them, as introduced by
this commit, in appreciation of an earlier invite.[1]

[1] https://github.com/protesilaos/denote/issues/117#issuecomment-1354372589