Document how to use tree-based file prompt on demand
This is my solution to a request made by Mirko Hernandez on the
possible use of the old Denote file prompt.  It is better not to
introduce a user option for this case, nor to keep multiple variants
of the 'denote-file-prompt' in denote.el, as we want to avoid
confusing the users.

Mirko's feedback was provided in issue 121 on the GitHub mirror:
<https://github.com/protesilaos/denote/issues/121>.