Add documentation to select silo interactively
I am providing this in response to a request from GitHub user
rbenit68.  The discussion took place in issue 127 on the GitHub
mirror, with the participation of Mirko Hernandez:
<https://github.com/protesilaos/denote/issues/127>.

The custom code I provide is the expanded version of an idea put forth
by Mirko, to whom I am thankful.