Make sure backlinks get correct 'denote-directory' path
The backlinks buffer will now get the correct path when it is
generated inside a silo.  This is related to issue 129 reported by
hapst3r on the GitHub mirror:
<https://github.com/protesilaos/denote/issues/129>.

This change is necessary because .dir-locals do not work for buffers,
so we must get the value from the file that calls 'denote-link-backlinks'.