Require the 'compat' library
This is for backward compatibility with previous Emacs versions.

I was already planning to incorporate 'compat' as I want to make
Denote available for Emacs 27.1.  Though now we need this
compatibility layer for commit 9d048ba.  Supporting older Emacs
versions requires some more work, but 'compat' gives us what we need
to make it happen.

Thanks to Alan Schmitt for informing me that
'file-name-parent-directory' is for Emacs 29.1:
<https://lists.sr.ht/~protesilaos/denote/%3C87fsb72nge.fsf%40protesilaos.com%3E#%3Cm2cz68ekng.fsf@eduroam-085c.sophia.inria.fr%3E>