REMOVE recursive search for silo root; keep simpler check
This was never a good idea but I wanted to see what the implications
are.  We do not need this anymore due to commit 3d2e1d5 which
simplifies 'denote--silo-p'.

Thanks to relict007, Hilde Rhyne, Mirko Hernández, Noboru
Ota (nobiot), Alan Schmitt, and hapst3r for their participation in the
relevant discussions:

- https://lists.sr.ht/~protesilaos/denote/%3C87fsb72nge.fsf%40protesilaos.com%3E
- https://lists.sr.ht/~protesilaos/denote/%3C80CBB671-D812-4EA8-8C80-85F9F4144051%40disroot.org%3E
- https://lists.sr.ht/~protesilaos/denote/%3C87pma6t59i.fsf%40kotlak.com%3E
- https://github.com/protesilaos/denote/issues/129 (GitHub mirror)