Implement sluggify scheme for signatures
Following the Denote file-naming scheme, we use the equals sign to
separate individual words in the SIGNATURE field.  Signatures are
anchored by the "==" field separator.  For context:

- Titles which are anchroned by the "--" field separator use hyphens.
- Keywords which are anchroned by the "__" field separator use underscores.

Thanks to Alfredo Borrás and Jeremy Friesen for discussiing this issue
on the mailing list:
<https://lists.sr.ht/~protesilaos/denote/%3C2A597B4E-5F18-4D97-9457-B3C859DAA020%40zoho.eu%3E>.

Thanks to Kai von Fintel for doing the same on the GitHub mirror in
issue 147: <https://github.com/protesilaos/denote/issues/147>.