Expose public functions for returning links or backlinks
We are doing the work in standalone functions that can be called from
Lisp, instead of limiting it to interactive functions.  The code is
more composable as a result.

This is in response to the mailing list thread started by relict007:
<https://lists.sr.ht/~protesilaos/denote/%3C87a5ygk6yi.fsf@kotlak.com%3E>

relict007 is the developer of the 'denote-cache' package (in
progress): <https://git.sr.ht/~relict007/denote-cache>.