Introduce variable `user-enforced-denote-directory`
The full context is captured on the mailing list[^1]. In brief, the
problem and proposed solution:

Problem  : Function `denote-directory` gives global preference to
           `.dir-locals.el` in all contexts, with no way to override
           it.

Solution : Allow an override so that users can create appropriate
           wrapper functions which work properly in all contexts.

This commit adds a new variable `user-enforced-denote-directory`,
which the end-user can use in their wrapper functions to reliably
control which silo their function should execute in.

This commit also updates the README and modifies the following
functions to demonstrate the use of the variable:

- `my-denote-pick-silo-then-command` (this is the base-case)
- `my-denote-org-extract-subtree` (to cut notes from org files)

The section `Use custom commands to select a silo` contains a small
explanation of the new variable.

[^1]:
https://lists.sr.ht/~protesilaos/denote/%3CCABzEscY9f12VDk64nE9OXDJAZMjsLUT2vrw-es%3D4yO%3Dh6ZJ5wQ%40mail.gmail.com%3E#%3CCABzEscY9f12VDk64nE9OXDJAZMjsLUT2vrw-es=4yO=h6ZJ5wQ@mail.gmail.com%3E