Make rename operations avoid duplicate identifiers
Before, it was possible for Denote to create duplicate identifiers
while renaming non-Denote files that had the same modification date.
Files can have the same modification date when they are processed by a
program, such as with Git or the 'touch' command.

What Denote now does is check for the presence of the given identifier
and produce a new one that is unique.  Depending on the command, the
check is done against the 'denote-directory' or the set of marked
Dired files.

The command 'denote-dired-rename-marked-files' gains a new optional
parameter, which is interactively passed by the double prefix
argument ('C-u C-u' with default key bindings).  I am not sure this is
the best interface, as we may prefer to always run the check for
duplicate identifiers.  The reason I am doing it this way with the
prefix argument is to avoid an unconditional potentially expensive
operation while renaming a large set of files.