Add functions to return file extension reliably (also with encryption)
We used to retrieve the 'file-name-extension' without checking if it
was a composite of the file type extension (e.g. ".org") and an
encryption extension like ".gpg".  Accounting for this scenario allows
us to properly support encrypted note files, whose extension is
something like ".org.gpg".

Check the two new public functions, 'denote-get-file-extension' and
'denote-get-file-extension-sans-encryption', for the implementation
details.

Thanks to Jens Östlund for reporting a bug with 'denote-keywords-add'
on an encrypted file, which prompted me to investigate this further.
This was done in issue 172 on the GitHub mirror:
<https://github.com/protesilaos/denote/issues/172>.

In short, we had a problem with all operations that needed to retrieve
the file extension when that included an encryption component.