Revert "Make 'denote-{title,signature}-prompt' complete from its history"
This reverts commit 0d855bbf1264347a16ce54d707af3698a4ec8eb6.

The idea to have completion is good, though this way it breaks with
the default minibuffer setup because SPC tries to complete the
inputted text. Thus the user cannot type freely. This does not happen
with packages that enhance the minibuffer User Interface, such as
'vertico'.

Thanks to Suhail Singh for reporting this bug:
<https://lists.sr.ht/~protesilaos/denote/%3C652d82c0.c80a0220.e6282.dc47%40mx.google.com%3E#%3C65392fa6.050a0220.da61c.0ac8@mx.google.com%3E>.