Re-enable 'denote-{title,signature}-prompt' history completion
I first implemented this in commit 0d855bb. However, it did not work
with the default minibuffer because the SPC key performs
completion (popping up the Completions buffer). So users could not
easily input an arbitrary string for the title/signature. I thus
reverted that commit in 9f692cb.

Stefan Monnier suggested the use of the 'minibuffer-with-setup-hook',
which lets us disable SPC completion for the purposes of these
functions. This is most welcome as the functionality is nice to have.

Stefan's feedback was provided on the emacs-devel mailing list:
<https://lists.gnu.org/archive/html/emacs-devel/2023-10/msg00631.html>.