Clarify wording about how multi-word keywords are handled
In short, you have to follow some convention that control the letter
casing of individual words to make them readable, while keeping them
part of the same keyword. Using the new user option we provide:

    (setq denote-file-name-letter-casing
          '((title . downcase)
            (signature . downcase)
            (keywords . verbatim) ; notice the `verbatim' here
            (t . downcase)))

With this in place, typing something like 'camelCase' at the keywords
prompt will insert that as-is and treat as a single keyword.

I am clarifying the documentation here as part of the process of
introducing the new 'denote-file-name-letter-casing' option. I do it
in response to the question raised by Peter Smith in issue 184 on the
GitHub mirror: <https://github.com/protesilaos/denote/issues/184>.