Stop using the 'denote--rename-file-subr' in 'denote-rename-file'
That abstraction was not good enough and thus makes things more
difficult for us. I am reverting to what we had before. We will see
how best to proceed.

Thanks to Jean-Philippe Gagné Guay for telling me about the drawbacks
of the 'denote--rename-file-subr'. This was done on the GitHub mirror.